fix: correct type cast error
Closes #6
Closes #8
