--- cgo/sha1_windows.go
@@ -1,22 +1,0 @@
-//go:build windows
-// +build windows
-
-package cgo
-
-// #include <sha1.h>
-// #include <stdlib.h>
-// #include <stddef.h>
-import "C"
-
-import "unsafe"
-
-func (d *digest) Write(p []byte) (nn int, err error) {
-	if len(p) == 0 {
-		return 0, nil
-	}
-
-	data := (*C.char)(unsafe.Pointer(&p[0]))
-	C.SHA1DCUpdate(&d.ctx, data, (C.ulonglong)(len(p)))
-
-	return len(p), nil
-}
